bugs fixed

+++ apis/table-delete.php
@@ -17,9 +17,11 @@ if (!isSuperAdmin($db, $userId, $tableName)) {
 	return;
 }
 
-$db->query("delete from data_requests where tableName='$tableName'");
-$db->query("delete from guest_permissions where tableName='$tableName'");
-$db->query("delete from table_admins where tableName='$tableName'");
+if (!doesTableExist($db, $tableName)) {
+	echo '{"status" : "failed", "message" : "No such table"}';
+	return;
+}
+
 $db->query("delete from tables_info where tableName='$tableName'");
 $db->query("drop table $tableName");
 echo '{"status" : "success"}';
